requirements.txt
accelerate==1.3.0
black==25.1.0
black[jupyter]==25.1.0
csvkit==2.0.1
datasets==3.2.0
decord==0.6.0
deepspeed==0.16.3
diffusers @ git+https://github.com/huggingface/diffusers.git@9a147b82f72e5df4553cb0f845bb957be3aa6028
flatten-dict==0.4.2
GitPython==3.1.44
huggingface_hub[cli]==0.27.0
hydra-core==1.3.2
imageio-ffmpeg==0.6.0
isort==6.0.1
opencv-python-headless==4.10.0.84
peft==0.14.0
pre_commit==4.1.0
protobuf==5.29.2
pytest==8.3.4
python-dotenv==1.0.1
seaborn==0.13.2
sentencepiece==0.2.0
tensorboard==2.18.0
transformers==4.48.3
